move autobind and debouncePromise to common/utils
