+++ src/common/utils/index.ts
@@ -1,7 +1,14 @@
-// Common utils (main/renderer)
+// Common utils (main OR renderer)
 
+export * from "./app-version"
+export * from "./autobind"
 export * from "./base64"
 export * from "./camelCase"
-export * from "./splitArray"
+export * from "./cloneJson"
+export * from "./debouncePromise"
+export * from "./defineGlobal"
 export * from "./getRandId"
+export * from "./splitArray"
+export * from "./saveToAppFiles"
+export * from "./singleton"
 export * from "./cloneJson"
